#1697: extract input-filter eval + filter-log emitters into filter.rs
Move the interface-input-filter evaluation and filter-log emission
helpers out of poll_descriptor/mod.rs into a new sibling module
poll_descriptor/filter.rs, with a per-function inline policy (NOT a
blanket #[inline(never)]) so the cheap common-case guards stay folded
into their hot/warm callers while only the rare/heavy bodies are forced
out of line:

- filter_log_ingress_zone_id / filter_log_egress_zone_id: #[inline]
  leaf helpers (called from both inline and cold callers).
- emit_cached_input_filter_log / emit_cached_output_filter_log are
  called UNCONDITIONALLY from stage_flow_cache_hit (#[inline(always)],
  the established-flow fast path). They stay #[inline] so the None
  filter-log guard folds into the fast path: the common no-logging case
  is a load + branch with no call and no 96-byte UserspaceDpMeta copy.
  The output emitter's rare non-None tail is split into a new
  #[cold] #[inline(never)] emit_cached_output_filter_log_tail; the input
  emitter's tail is already just a call to the cold
  emit_input_filter_log_match. This is the round-1 PLAN-KILL fix: v1
  would have marked these #[inline(never)] and forced two unconditional
  per-packet calls plus two 96-byte struct copies on every cache hit.
- evaluate_dscp_sensitive_input_filter_on_session_hit runs per packet on
  the session-hit path for DSCP-sensitive-filter configs (those are not
  flow-cached, flow_cache.rs:285). It stays #[inline] so the cheap
  interface_input_filter_has_dscp_match guard folds in and returns None
  with no call when no DSCP filter is configured.
- evaluate_non_pbr_input_filter / _log_only, emit_input_filter_log_match,
  apply_lo0_filter_action: #[cold] #[inline(never)] exception bodies for
  .text.unlikely placement away from the hot loop's cache lines.

flow_cache_hit.rs now imports emit_cached_input_filter_log /
emit_cached_output_filter_log from super::filter. mod.rs imports the
helpers it still calls from filter. Bodies are byte-identical to their
previous location apart from the inline attributes and the single tail
split.

mod.rs drops from 2983 to 2437 LOC.

Validation: cargo build clean (no new warnings).
